[OF-1688] refac: renamed and moved 'Status/Status.h/cpp' to 'ServiceStatus.h/cpp'
